pyproject.toml
[tool.poetry]
name = "streaming-video-processing"
version = "0.1.0"
description = ""
authors = ["Felipe Allegretti <felipe@allegretti.me>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
confluent-kafka = "2.0.2"
pymongo = "4.3.3"
opencv-python = "4.7.0.72"
tensorflow = "2.11.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"

[tool.black]
line-length = 120
verbose = true
preview = true

[tool.isort]
profile = "black"
